daemon/volumes: move []container.Mount sorting into container pkg
