+++ daemon/volumes_unix.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -79,7 +78,7 @@ func (daemon *Daemon) setupMounts(c *container.Container) ([]container.Mount, er
 		}
 	}
 
-	mounts = sortMounts(mounts)
+	mounts = container.SortMounts(mounts)
 	netMounts := c.NetworkMounts()
 	// if we are going to mount any of the network files from container
 	// metadata, the ownership must be set properly for potential container
@@ -91,14 +90,6 @@ func (daemon *Daemon) setupMounts(c *container.Container) ([]container.Mount, er
 		}
 	}
 	return append(mounts, netMounts...), nil
-}
-
-// sortMounts sorts an array of mounts in lexicographic order. This ensure that
-// when mounting, the mounts don't shadow other mounts. For example, if mounting
-// /etc and /etc/resolv.conf, /etc/resolv.conf must not be mounted first.
-func sortMounts(m []container.Mount) []container.Mount {
-	sort.Sort(mounts(m))
-	return m
 }
 
 // setBindModeIfNull is platform specific processing to ensure the
